+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneWifiConfigurationPolicyAndroidOpenSourceProjectBasic/MSFT_IntuneWifiConfigurationPolicyAndroidOpenSourceProjectBasic.psm1
@@ -26,15 +26,47 @@ function Get-TargetResource
         $ConnectWhenNetworkNameIsHidden,
 
         [Parameter()]
+        [System.Boolean]
+        $ForcePreSharedKeyUpdate,
+
+        [Parameter()]
         [System.String]
         $NetworkName,
+
+        [Parameter()]
+        [System.String]
+        $PreSharedKey,
+
+        [Parameter()]
+        [System.Boolean]
+        $PreSharedKeyIsSet,
+
+        [Parameter()]
+        [System.String]
+        $ProxyAutomaticConfigurationUrl,
+
+        [Parameter()]
+        [System.String[]]
+        $ProxyExclusionList,
+
+        [Parameter()]
+        [System.String]
+        $ProxyManualAddress,
+
+        [Parameter()]
+        [System.Int32]
+        $ProxyManualPort,
+
+        [Parameter()]
+        [ValidateSet('none', 'manual', 'automatic')]
+        $ProxySetting,
 
         [Parameter()]
         [System.String]
         $Ssid,
 
         [Parameter()]
-        [ValidateSet('open', 'wpaEnterprise', 'wpa2Enterprise')]
+        [ValidateSet('open', 'wep', 'wpaPersonal')]
         [System.String]
         $WiFiSecurityType,
 
@@ -102,12 +134,13 @@ function Get-TargetResource
         $getValue = Get-MgBetaDeviceManagementDeviceConfiguration -DeviceConfigurationId $id -ErrorAction SilentlyContinue
 
         #region resource generator code
-        if ($null -ne $getValue)
+        if ($null -eq $getValue)
         {
             $getValue = Get-MgBetaDeviceManagementDeviceConfiguration -Filter "DisplayName eq '$Displayname'" -ErrorAction SilentlyContinue | Where-Object `
             -FilterScript { `
-                $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.androidWorkProfileWiFiConfiguration' `
+                $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.aospDeviceOwnerWiFiConfiguration' `
             }
+            $Id = $getValue.Id
         }
         #endregion
 
@@ -126,6 +159,13 @@ function Get-TargetResource
             ConnectAutomatically           = $getValue.AdditionalProperties.connectAutomatically
             ConnectWhenNetworkNameIsHidden = $getValue.AdditionalProperties.connectWhenNetworkNameIsHidden
             NetworkName                    = $getValue.AdditionalProperties.networkName
+            PreSharedKey                   = $getValue.AdditionalProperties.preSharedKey
+            PreSharedKeyIsSet              = $getValue.AdditionalProperties.preSharedKeyIsSet
+            ProxyAutomaticConfigurationUrl = $getValue.AdditionalProperties.proxyAutomaticConfigurationUrl
+            ProxyExclusionList             = $getValue.AdditionalProperties.proxyExclusionList
+            ProxyManualAddress             = $getValue.AdditionalProperties.proxyManualAddress
+            ProxyManualPort                = $getValue.AdditionalProperties.proxyManualPort
+            ProxySetting                   = $getValue.AdditionalProperties.proxySetting
             Ssid                           = $getValue.AdditionalProperties.ssid
             WiFiSecurityType               = $getValue.AdditionalProperties.wiFiSecurityType
             Ensure                         = 'Present'
@@ -134,7 +174,7 @@ function Get-TargetResource
             TenantId                       = $TenantId
             ApplicationSecret              = $ApplicationSecret
             CertificateThumbprint          = $CertificateThumbprint
-            Managedidentity                = $ManagedIdentity.IsPresent
+            ManagedIdentity                = $ManagedIdentity.IsPresent
         }
 
         $assignmentsValues = Get-MgBetaDeviceManagementDeviceConfigurationAssignment -DeviceConfigurationId $Id
@@ -193,15 +233,47 @@ function Set-TargetResource
         $ConnectWhenNetworkNameIsHidden,
 
         [Parameter()]
+        [System.Boolean]
+        $ForcePreSharedKeyUpdate,
+
+        [Parameter()]
         [System.String]
         $NetworkName,
+
+        [Parameter()]
+        [System.String]
+        $PreSharedKey,
+
+        [Parameter()]
+        [System.Boolean]
+        $PreSharedKeyIsSet,
+
+        [Parameter()]
+        [System.String]
+        $ProxyAutomaticConfigurationUrl,
+
+        [Parameter()]
+        [System.String[]]
+        $ProxyExclusionList,
+
+        [Parameter()]
+        [System.String]
+        $ProxyManualAddress,
+
+        [Parameter()]
+        [System.Int32]
+        $ProxyManualPort,
+
+        [Parameter()]
+        [ValidateSet('none', 'manual', 'automatic')]
+        $ProxySetting,
 
         [Parameter()]
         [System.String]
         $Ssid,
 
         [Parameter()]
-        [ValidateSet('open', 'wpaEnterprise', 'wpa2Enterprise')]
+        [ValidateSet('open', 'wep', 'wpaPersonal')]
         [System.String]
         $WiFiSecurityType,
 
@@ -240,6 +312,14 @@ function Set-TargetResource
         $ManagedIdentity
     )
 
+    if ($ProxySetting -ne 'automatic' -and $ProxyAutomaticConfigurationUrl -ne '') {
+        throw 'ProxyAutomaticConfigurationUrl must be empty if ProxySetting is not "automatic"'
+    }
+
+    if ($PreSharedKeyIsSet -and $PreSharedKey -eq '') {
+        throw 'PreSharedKey is required but was not set.'
+    }
+
     try
     {
         $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' `
@@ -264,6 +344,7 @@ function Set-TargetResource
 
     $currentInstance = Get-TargetResource @PSBoundParameters
 
+    $PSBoundParameters.Remove('ForcePreSharedKeyUpdate') | Out-Null
     $PSBoundParameters.Remove('Ensure') | Out-Null
     $PSBoundParameters.Remove('Credential') | Out-Null
     $PSBoundParameters.Remove('ApplicationId') | Out-Null
@@ -303,6 +384,9 @@ function Set-TargetResource
 
         if ($AdditionalProperties)
         {
+            if ($AdditionalProperties['proxyAutomaticConfigurationUrl'] -eq '') {
+                $AdditionalProperties['proxyAutomaticConfigurationUrl'] = $null
+            }
             $CreateParameters.add('AdditionalProperties', $AdditionalProperties)
         }
 
@@ -353,6 +437,9 @@ function Set-TargetResource
 
         if ($AdditionalProperties)
         {
+            if ($AdditionalProperties['proxyAutomaticConfigurationUrl'] -eq '') {
+                $AdditionalProperties['proxyAutomaticConfigurationUrl'] = $null
+            }
             $UpdateParameters.add('AdditionalProperties', $AdditionalProperties)
         }
 
@@ -373,7 +460,6 @@ function Set-TargetResource
     elseif ($Ensure -eq 'Absent' -and $currentInstance.Ensure -eq 'Present')
     {
         Write-Verbose -Message "Removing {$DisplayName}"
-
         #region resource generator code
         Remove-MgBetaDeviceManagementDeviceConfiguration -DeviceConfigurationId $currentInstance.Id
         #endregion
@@ -408,15 +494,47 @@ function Test-TargetResource
         $ConnectWhenNetworkNameIsHidden,
 
         [Parameter()]
+        [System.Boolean]
+        $ForcePreSharedKeyUpdate,
+
+        [Parameter()]
         [System.String]
         $NetworkName,
+
+        [Parameter()]
+        [System.String]
+        $PreSharedKey,
+
+        [Parameter()]
+        [System.Boolean]
+        $PreSharedKeyIsSet,
+
+        [Parameter()]
+        [System.String]
+        $ProxyAutomaticConfigurationUrl,
+
+        [Parameter()]
+        [System.String[]]
+        $ProxyExclusionList,
+
+        [Parameter()]
+        [System.String]
+        $ProxyManualAddress,
+
+        [Parameter()]
+        [System.Int32]
+        $ProxyManualPort,
+
+        [Parameter()]
+        [ValidateSet('none', 'manual', 'automatic')]
+        $ProxySetting,
 
         [Parameter()]
         [System.String]
         $Ssid,
 
         [Parameter()]
-        [ValidateSet('open', 'wpaEnterprise', 'wpa2Enterprise')]
+        [ValidateSet('open', 'wep', 'wpaPersonal')]
         [System.String]
         $WiFiSecurityType,
 
@@ -455,6 +573,20 @@ function Test-TargetResource
         $ManagedIdentity
     )
 
+    if ($ProxySetting -ne 'automatic' -and $ProxyAutomaticConfigurationUrl -ne '') {
+        throw 'ProxyAutomaticConfigurationUrl must be empty if ProxySetting is not "automatic".'
+    }
+
+    if ($PreSharedKeyIsSet -and $PreSharedKey -eq '') {
+        throw 'PreSharedKey is required but was not set.'
+    }
+
+    if ($ForcePreSharedKeyUpdate) {
+        Write-Verbose -Message "ForcePreSharedKeyUpdate was set, always force an update."
+        Write-Verbose -Message "Test-TargetResource returned $false"
+        return $false
+    }
+
     #Ensure the proper dependencies are installed in the current environment.
     Confirm-M365DSCDependencies
 
@@ -472,7 +604,7 @@ function Test-TargetResource
     $CurrentValues = Get-TargetResource @PSBoundParameters
     $ValuesToCheck = ([Hashtable]$PSBoundParameters).clone()
 
-    if ($CurrentValues.Ensure -eq 'Absent')
+    if ($CurrentValues.Ensure -ne $PSBoundParameters.Ensure)
     {
         Write-Verbose -Message "Test-TargetResource returned $false"
         return $false
@@ -517,6 +649,8 @@ function Test-TargetResource
         }
     }
 
+    $ValuesToCheck.Remove('ForcePreSharedKeyUpdate') | Out-Null
+    $ValuesToCheck.Remove('PreSharedKey') | Out-Null
     $ValuesToCheck.Remove('Credential') | Out-Null
     $ValuesToCheck.Remove('ApplicationId') | Out-Null
     $ValuesToCheck.Remove('TenantId') | Out-Null
@@ -601,7 +735,7 @@ function Export-TargetResource
         [array]$getValue = Get-MgBetaDeviceManagementDeviceConfiguration `
             -ErrorAction Stop | Where-Object `
             -FilterScript { `
-                $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.androidWorkProfileEnterpriseWiFiConfiguration'  `
+                $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.aospDeviceOwnerWiFiConfiguration'  `
         }
         #endregion
 
@@ -627,7 +761,7 @@ function Export-TargetResource
                 TenantId              = $TenantId
                 ApplicationSecret     = $ApplicationSecret
                 CertificateThumbprint = $CertificateThumbprint
-                Managedidentity       = $ManagedIdentity.IsPresent
+                ManagedIdentity       = $ManagedIdentity.IsPresent
             }
 
             $Results = Get-TargetResource @Params
@@ -691,9 +825,6 @@ function Export-TargetResource
         return ''
     }
 }
-
-
-
 function Get-M365DSCAdditionalProperties
 {
     [CmdletBinding()]
@@ -709,11 +840,18 @@ function Get-M365DSCAdditionalProperties
         'ConnectAutomatically'
         'ConnectWhenNetworkNameIsHidden'
         'NetworkName'
+        'PreSharedKey'
+        'PreSharedKeyIsSet'
+        'ProxyAutomaticConfigurationUrl'
+        'ProxyExclusionList'
+        'ProxyManualAddress'
+        'ProxyManualPort'
+        'ProxySetting'
         'Ssid'
         'WiFiSecurityType'
     )
 
-    $results = @{'@odata.type' = '#microsoft.graph.androidWorkProfileEnterpriseWiFiConfiguration' }
+    $results = @{'@odata.type' = '#microsoft.graph.aospDeviceOwnerWiFiConfiguration' }
     $cloneProperties = $Properties.clone()
     foreach ($property in $cloneProperties.Keys)
     {
@@ -744,6 +882,7 @@ function Get-M365DSCAdditionalProperties
             }
 
             $results.Add($propertyName, $propertyValue)
+
         }
     }
     if ($results.Count -eq 1)