+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneWifiConfigurationPolicyAndroidEnterpriseWorkProfileBasic.Tests.ps1
@@ -15,7 +15,7 @@ Import-Module -Name (Join-Path -Path $M365DSCTestFolder `
         -Resolve)
 
 $Global:DscHelper = New-M365DscUnitTestHelper -StubModule $CmdletModule `
-    -DscResource 'IntuneWifiConfigurationPolicyAndroidOpenSourceProject' -GenericStubModule $GenericStubPath
+    -DscResource 'IntuneWifiConfigurationPolicyAndroidEnterpriseWorkProfileBasic' -GenericStubModule $GenericStubPath
 Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
     InModuleScope -ModuleName $Global:DscHelper.ModuleName -ScriptBlock {
         Invoke-Command -ScriptBlock $Global:DscHelper.InitializeScript -NoNewScope
@@ -54,7 +54,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         }
 
         # Test contexts
-        Context -Name 'The IntuneWifiConfigurationPolicyAndroidOpenSourceProject should exist but it DOES NOT' -Fixture {
+        Context -Name 'The IntuneWifiConfigurationPolicyAndroidEnterpriseWorkProfileBasic should exist but it DOES NOT' -Fixture {
             BeforeAll {
                 $testParams = @{
                     ConnectAutomatically           = $True
@@ -63,8 +63,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     DisplayName                    = 'FakeStringValue'
                     Id                             = 'FakeStringValue'
                     NetworkName                    = 'FakeStringValue'
-                    PreSharedKey                   = 'FakeStringValue'
-                    PreSharedKeyIsSet              = $True
                     Ssid                           = 'FakeStringValue'
                     WiFiSecurityType               = 'open'
 
@@ -88,7 +86,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             }
         }
 
-        Context -Name 'The IntuneWifiConfigurationPolicyAndroidOpenSourceProject exists but it SHOULD NOT' -Fixture {
+        Context -Name 'The IntuneWifiConfigurationPolicyAndroidEnterpriseWorkProfileBasic exists but it SHOULD NOT' -Fixture {
             BeforeAll {
                 $testParams = @{
                     ConnectAutomatically           = $True
@@ -97,8 +95,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     DisplayName                    = 'FakeStringValue'
                     Id                             = 'FakeStringValue'
                     NetworkName                    = 'FakeStringValue'
-                    PreSharedKey                   = 'FakeStringValue'
-                    PreSharedKeyIsSet              = $True
                     Ssid                           = 'FakeStringValue'
                     WiFiSecurityType               = 'open'
 
@@ -109,14 +105,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 Mock -CommandName Get-MgBetaDeviceManagementDeviceConfiguration -MockWith {
                     return @{
                         AdditionalProperties = @{
-                            '@odata.type'                  = '#microsoft.graph.aospDeviceOwnerWifiConfiguration'
+                            '@odata.type'                  = '#microsoft.graph.androidWorkProfileWifiConfiguration'
                             NetworkName                    = 'FakeStringValue'
                             WiFiSecurityType               = 'open'
                             ConnectAutomatically           = $True
-                            PreSharedKey                   = 'FakeStringValue'
                             ConnectWhenNetworkNameIsHidden = $True
                             Ssid                           = 'FakeStringValue'
-                            PreSharedKeyIsSet              = $True
 
                         }
                         Description          = 'FakeStringValue'
@@ -140,7 +134,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 Should -Invoke -CommandName Remove-MgBetaDeviceManagementDeviceConfiguration -Exactly 1
             }
         }
-        Context -Name 'The IntuneWifiConfigurationPolicyAndroidOpenSourceProject Exists and Values are already in the desired state' -Fixture {
+        Context -Name 'The IntuneWifiConfigurationPolicyAndroidEnterpriseWorkProfileBasic Exists and Values are already in the desired state' -Fixture {
             BeforeAll {
                 $testParams = @{
                     ConnectAutomatically           = $True
@@ -149,8 +143,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     DisplayName                    = 'FakeStringValue'
                     Id                             = 'FakeStringValue'
                     NetworkName                    = 'FakeStringValue'
-                    PreSharedKey                   = 'FakeStringValue'
-                    PreSharedKeyIsSet              = $True
                     Ssid                           = 'FakeStringValue'
                     WiFiSecurityType               = 'open'
 
@@ -161,14 +153,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 Mock -CommandName Get-MgBetaDeviceManagementDeviceConfiguration -MockWith {
                     return @{
                         AdditionalProperties = @{
-                            '@odata.type'                  = '#microsoft.graph.aospDeviceOwnerWifiConfiguration'
+                            '@odata.type'                  = '#microsoft.graph.androidWorkProfileWifiConfiguration'
                             NetworkName                    = 'FakeStringValue'
                             WiFiSecurityType               = 'open'
                             ConnectAutomatically           = $True
-                            PreSharedKey                   = 'FakeStringValue'
                             ConnectWhenNetworkNameIsHidden = $True
                             Ssid                           = 'FakeStringValue'
-                            PreSharedKeyIsSet              = $True
 
                         }
                         Description          = 'FakeStringValue'
@@ -185,7 +175,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             }
         }
 
-        Context -Name 'The IntuneWifiConfigurationPolicyAndroidOpenSourceProject exists and values are NOT in the desired state' -Fixture {
+        Context -Name 'The IntuneWifiConfigurationPolicyAndroidEnterpriseWorkProfileBasic exists and values are NOT in the desired state' -Fixture {
             BeforeAll {
                 $testParams = @{
                     ConnectAutomatically           = $True
@@ -194,8 +184,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     DisplayName                    = 'FakeStringValue'
                     Id                             = 'FakeStringValue'
                     NetworkName                    = 'FakeStringValue'
-                    PreSharedKey                   = 'FakeStringValue'
-                    PreSharedKeyIsSet              = $True
                     Ssid                           = 'FakeStringValue'
                     WiFiSecurityType               = 'open'
 
@@ -209,8 +197,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                             Ssid             = 'FakeStringValue'
                             NetworkName      = 'FakeStringValue'
                             WiFiSecurityType = 'open'
-                            PreSharedKey     = 'FakeStringValue'
-                            '@odata.type'    = '#microsoft.graph.aospDeviceOwnerWifiConfiguration'
+                            '@odata.type'    = '#microsoft.graph.androidWorkProfileWifiConfiguration'
 
                         }
                         Description          = 'FakeStringValue'
@@ -237,6 +224,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
         Context -Name 'ReverseDSC Tests' -Fixture {
             BeforeAll {
+                $Global:PartialExportFileName = "$(New-Guid).partial.ps1"
                 $testParams = @{
                     Credential = $Credential
                 }
@@ -244,14 +232,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 Mock -CommandName Get-MgBetaDeviceManagementDeviceConfiguration -MockWith {
                     return @{
                         AdditionalProperties = @{
-                            '@odata.type'                  = '#microsoft.graph.aospDeviceOwnerWifiConfiguration'
+                            '@odata.type'                  = '#microsoft.graph.androidWorkProfileEnterpriseWiFiConfiguration'
                             NetworkName                    = 'FakeStringValue'
                             WiFiSecurityType               = 'open'
                             ConnectAutomatically           = $True
-                            PreSharedKey                   = 'FakeStringValue'
                             ConnectWhenNetworkNameIsHidden = $True
                             Ssid                           = 'FakeStringValue'
-                            PreSharedKeyIsSet              = $True
 
                         }
                         Description          = 'FakeStringValue'